Fix the align argument of the memory instructions
The align argument should be an exponential value.
If there is no align argument, it should be the default value.

+++ wain-validate/src/error.rs
@@ -25,8 +25,8 @@ pub enum ErrorKind {
         idx: u32,
     },
     TooLargeAlign {
-        align: u8,
-        bits: u8,
+        align: u32,
+        bits: u32,
     },
     InvalidLimitRange(u32, u32),
     LimitsOutOfRange {
